+++ js/data.js
@@ -1,102 +1,5 @@
-//Тест на переполнение формы:
+import { getRandomInt, getUniqueId } from './util.js';
 
-/*function stringLength(string, maxLength) {
-  return string.length <= maxLength;
-}
-
-console.log(stringLength(10, 5));
-
-
-//Тест на определение полиндрома:
-
-function isItPalindrome(string) {
-
-  let stringNoSpaces = '';
-
-  for (let i = 0; i < string.length; i++) {
-    if (string[i] !== ' ') {
-      stringNoSpaces += string[i];
-    }
-  }
-
-  stringNoSpaces = stringNoSpaces.toLowerCase();
-
-  let stringNoSpacesRevert = '';
-
-  for (let i = (stringNoSpaces.length - 1); i >= 0; i--) {
-    stringNoSpacesRevert += stringNoSpaces[i];
-  }
-
-  return stringNoSpaces === stringNoSpacesRevert;
-}
-
-console.log(isItPalindrome('Анана'));
-
-
-//Дополнительное задание:
-
-function wtf(string) {
-
-  let result = '';
-
-  for (let i = 0; i < string.length; i++) {
-    if (string[i] != ' ') {
-      let isItNumber = +string[i];
-      if (Number.isInteger(isItNumber)) {
-        let absoluteNumber = Math.abs(isItNumber);
-        result += absoluteNumber;
-      }
-    }
-  }
-
-  if (result === '') {
-    return NaN;
-  }
-
-  return result;
-
-}
-
-console.log(wtf('Сегодня 20 февраля 2024 года, за окном -2.5 градуса. Агент 007 провалил задание, т.к. не понял как по-другому вернуть NaN при нечисловой строке'));
-
-
-//Вариант решения задания с тренажёра:
-
-let qualificationDistance = 200;
-let attempts = [120, 150, 160, 201, 203, 180, 202];
-let qualified = false;
-let averageBest = 0;
-let attemptQuantitys = 3;
-
-
-for (let i = 0; i <= attempts.length - 2; i++) {
-  let bestAttempt = attempts[i];
-
-  for (let j = i + 1; j <= attempts.length - 1; j++){
-    if (bestAttempt > attempts[j]) {
-      bestAttempt = attempts[j];
-      let swap = attempts[i];
-      attempts[i] = bestAttempt;
-      attempts[j] = swap;
-    }
-  }
-}
-
-console.log(attempts);
-
-let attemptsSum = 0;
-
-for (let i = 1; i <= attemptQuantitys; i++) {
-  attemptsSum += attempts[attempts.length - i]
-}
-
-averageBest = attemptsSum / attemptQuantitys;
-
-if (averageBest > qualificationDistance) {
-  qualified = true;
-}
-*/
-//============================================
 
 const photoDescriptions = [
   'Коллекция морских ракушек',
@@ -155,34 +58,10 @@ const commentNames = [
   'Бобр Добр'
 ];
 
-/*Функция генерации случайных чисел*/
-
-const getRandomInt = (minInt, maxInt) => {
-  const randomInt = Math.floor(Math.random() * (maxInt - minInt + 1) + minInt);
-  return randomInt;
-};
-
-/*Функция генерации уникальных чисел в заданном диапазоне + проверка на уникальность*/
-
-const getUniqueId = (minInt, maxInt) => {
-  const uniqueIds = {};
-
-  return () => {
-    let uniqueId = getRandomInt(minInt, maxInt);
-    if (uniqueIds[uniqueId] === true) {
-      uniqueId = getRandomInt(minInt, maxInt);
-    }
-    uniqueIds[uniqueId] = true;
-
-    return uniqueId;
-  };
-};
-
 /*Генерация комментария*/
 
 const commentIdMinValue = 1;
 const commentIdMaxValue = 30;
-
 const commentUniqueId = getUniqueId(commentIdMinValue, commentIdMaxValue);
 
 const getComment = () => {
@@ -204,7 +83,7 @@ const getComment = () => {
 const getCommentsArr = () => {
   const commentsArr = [];
   const commentsCount = getRandomInt(0, 30);
-  while (commentsArr.length <= commentsCount) {
+  while (commentsArr.length < commentsCount) {
     commentsArr.push(getComment());
   }
 
@@ -222,12 +101,8 @@ const urlIdMaxValue = 25;
 const urlUniqueId = getUniqueId(urlIdMinValue, urlIdMaxValue);
 
 const getPhoto = () => {
-
   const id = photoUniqueId();
-
-
   const url = `photos/${urlUniqueId()}.jpg`;
-
   const description = photoDescriptions[getRandomInt(0, photoDescriptions.length - 1)];
   const likes = getRandomInt(15, 250);
   const comments = getCommentsArr();
@@ -252,9 +127,4 @@ const getPhotosArr = () => {
   return photosArr;
 };
 
-console.log(getPhotosArr());
-
-
-
-
-
+export { getPhotosArr };
